+++ docs/changes.rst
@@ -30,6 +30,15 @@ v1.7 (in development)
 
 - ``offset`` is now optional in :meth:`Link.from_track`.
 
+- Remove undocumented/internal method
+  :meth:`spotify.manager.SpotifySessionManager.wake`.
+  :meth:`spotify.manager.SpotifySessionManager.notify_main_thread` does the
+  same job.
+
+- Remove undocumented/internal method
+  :meth:`spotify.manager.SpotifySessionManager.terminate`. Use
+  :meth:`spotify.manager.SpotifySessionManager.disconnect` instead.
+
 **New features**
 
 - Added method :meth:`spotify.Playlist.owner`.